Add sub-tasks check-any-warnings, check-warnings and clear-warnings
The task continue:check-any-warnings does the opposite of
continue:continue:fail-on-warning. It fails, if there was no warning
issued.

The multi-task continue:check-warnings fails, if there was no warning
issued, or if there were other warnings issued, than specified for the
task.

The task continue:clear-warnings clears the warnings, which have been
remembered in the continuation blocks so far.